% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{alphanumeric_tbl}
\alias{alphanumeric_tbl}
\title{Expand an alphanumeric range into a tibble}
\usage{
alphanumeric_tbl(alphanumeric_range)
}
\arguments{
\item{alphanumeric_range}{A character string specifying an alphanumeric range using \code{:} notation}
}
\value{
A \link[tibble]{tibble} with row, col
}
\description{
Expand a tibble with row and col from an alphanumeric range with the \code{:} notation, e.g. A1:A3
}
\examples{
alphanumeric_tbl("A1:C3")
}
\concept{helper}
