% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as_factor}
\alias{as_factor}
\title{Convert features into factors}
\usage{
as_factor(tbl, con, drop_unused_levels = FALSE)
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} with character feature columns to be converted into factors}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}

\item{drop_unused_levels}{Boolean indicating whether to drop unused levels}
}
\value{
A \link[tibble]{tibble} with factor features
}
\description{
Convert character features into unordered or ordered factors.
}
\concept{transforming}
