% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as_flextable}
\alias{as_flextable}
\title{Convert tibble into flextable}
\usage{
as_flextable(tbl, autofit_part = "header", save = FALSE, filename = NULL)
}
\arguments{
\item{tbl}{A \link[tibble]{tibble}}

\item{autofit_part}{Character string specifying how to fit column widths ("all", "body", "header" or "footer")}

\item{save}{Boolean indicating whether to save the flextable}

\item{filename}{Character string specifying the filename}
}
\value{
A \link[flextable]{flextable}
}
\description{
Convert a tibble into a flextable with custom formatting.
}
\concept{helper}
