% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate to API}
\usage{
authenticate(req, oal = NULL)
}
\arguments{
\item{req}{A \link[httr2]{request}}

\item{oal}{An OAuth list return by \code{\link[=oauth]{oauth()}} with a \link[httr2]{oauth_client}, the auth_url, the redirect_uri and the scope}
}
\value{
A modified HTTP \link[httr2]{request} that will use OAuth;
}
\description{
\code{authenticate} is a wrapper around \link[httr2]{req_oauth_auth_code}.
It receives a \link[httr2]{request} and a OAuth list, returning a modified request.
}
\concept{connecting}
