% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{average_response}
\alias{average_response}
\title{Calculate average drug response}
\usage{
average_response(
  tbl,
  ...,
  by = c("drug_id", "assay_id", "sid", "pid"),
  FUN = mean,
  exclude_failed = FALSE
)
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} with at least the columns
\itemize{
\item pid
\item sid
\item drug_name
\item assay_failed
\item and the metrics selected in ...
}}

\item{...}{Tidyselect columns (metrics) to average over}

\item{by}{Variables to use for grouping}

\item{FUN}{Function to use calculate_zscore_from_stats for averaging}

\item{exclude_failed}{Boolean indicating whether to exclude assays that have failed QC (if \code{TRUE} then only \code{assay_failed=FALSE/NA})}
}
\value{
A \link[tibble]{tibble} with individual drug responses replaced by the averaged drug response
}
\description{
Calculate the average (mean or median) response across grouping variables (e.g. Idarubicin responses present on multiple plates) .
By default computes the mean response for each patient, sample, assays and drug across all plates.
}
\concept{scoring}
