% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_absolute_logEC}
\alias{calculate_absolute_logEC}
\title{Calculate absolute logEC50}
\usage{
calculate_absolute_logEC(effect, b, c, d, logEC50)
}
\arguments{
\item{effect}{Percent absolute effect (e.g. 50 or 20)}

\item{b}{Slope parameter}

\item{c}{Lower asymptote}

\item{d}{Upper asymptote}

\item{logEC50}{Logarithmic half-maximum effective concentration (relative EC50)}
}
\value{
Concentration at which the given response is observed
}
\description{
\code{calculate_absolute_logEC} computes the absolute effect concentration (e.g. absolute logEC50 or logEC20) for the LL.2, LL.3 or LL.4 model.
The absolute logEC50 is the concentration where the four-parameter log-logistic function crosses the 50\% effect treshold.
Note: logEC20 -> 20\% effect = 20\% inhibition = 80\% viability
}
\note{
This is equivalent to \code{drc::ED(mdl, effect/100, type="absolute")} where \code{mdl} is a \code{drc::drm()} object
}
\concept{scoring}
