% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_avg_dev}
\alias{calculate_avg_dev}
\title{Calculate the average replicate deviation}
\usage{
calculate_avg_dev(con, .sample_drug_mask = TRUE, readout_id = c(1, 5))
}
\arguments{
\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}

\item{.sample_drug_mask}{<\code{\link[rlang:args_data_masking]{data-masking}}> expression returning a logical vector and selecting the samples/assays and drugs}

\item{readout_id}{Vector of integers specifying the readout IDs of which the average deviation is calculated}
}
\value{
A \link[tibble]{tibble} with the averaged replicate deviations for all assays/samples
}
\description{
Calculate the average deviation of the replicates across the dilutions series.
}
\concept{scoring}
