% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_dmso_recovery}
\alias{calculate_dmso_recovery}
\title{Calculate DMSO recovery}
\usage{
calculate_dmso_recovery(tbl, con, readout_id = c(1, 5))
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} with at least the following columns:
\itemize{
\item sid
\item assay_id
\item readout_id (optional)
\item drug_id
\item exclude
}}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}

\item{readout_id}{Vector of integers specifying the readout IDs of which the DMSO recovery is calculated}
}
\value{
A \link[tibble]{tibble} with the plate averaged recoveries for all assays/samples
}
\description{
The DMSO recovery describes the ratio of cancer cells recovered after 96h (72h of drug incubation)
in the DMSO negative control identified by CyQuant with respect to the number of originally seeded cells
assessed by Trypan blue as specified in the DRP database in \code{assay->cancer_seeded}.
}
\concept{scoring}
