% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_dss}
\alias{calculate_dss}
\title{Calculate the drug sensitivity score (DSS)}
\usage{
calculate_dss(
  type,
  logAUC,
  t = 0.1,
  c_min = -1,
  c_max = 4,
  x1 = -1,
  x2 = 4,
  d = 1
)
}
\arguments{
\item{type}{DSS type}

\item{logAUC}{Logarithmic area-under-the-curve}

\item{t}{Minimum activity treshold}

\item{c_min}{Logarithmic lowest tested concentration}

\item{c_max}{Logarithmic highest tested concentration}

\item{x1}{Logarithmic lower integration limit}

\item{x2}{Logarithmic lower integration limit}

\item{d}{Upper asymptote}
}
\value{
Drug sensitivity score
}
\description{
\code{calculate_dss} computes the drug sensitivity score (DSS) from the logAUC within the lower/upper integration limits
}
\concept{scoring}
