% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_logAUC}
\alias{calculate_logAUC}
\title{Calculate the logarithmic area-under-curve (logAUC)}
\usage{
calculate_logAUC(b, c, d, logEC50, x1, x2)
}
\arguments{
\item{b}{Inverse slope parameter}

\item{c}{Lower asymptote}

\item{d}{Upper asymptote}

\item{logEC50}{Logarithmic half-maximum effective concentration (relative EC50)}

\item{x1}{Numeric indicating the minimal tested logarithmic concentration (lower integration limit)}

\item{x2}{Numeric indicating the maximal tested logarithmic concentration (upper integration limit)}
}
\value{
Area under the logarithmic dose response curve
}
\description{
\code{calculate_logAUC} computes the area-under-the-curve for the ll.2, ll.3 and ll.4 model
An analytical solution for the ll.5 model is currently not implemented.
}
\concept{scoring}
