% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_norm_logAUC}
\alias{calculate_norm_logAUC}
\title{Calculate the normalized area-under-curve (norm. logAUC)}
\usage{
calculate_norm_logAUC(logAUC, x1, x2)
}
\arguments{
\item{logAUC}{Logarithmic area-under-the-curve}

\item{x1}{Numeric indicating the minimal tested logarithmic concentration (lower integration limit)}

\item{x2}{Numeric indicating the maximal tested logarithmic concentration (upper integration limit)}
}
\value{
Normalized area under the logarithmic dose response curve
}
\description{
\code{calculate_norm_logAUC} computes the normalized AUC from the logAUC within the lower/upper integration limits
}
\concept{scoring}
