% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_norm_logEC50}
\alias{calculate_norm_logEC50}
\title{Calculate the normalized logEC50}
\usage{
calculate_norm_logEC50(logEC50, x1, x2, log_extra)
}
\arguments{
\item{logEC50}{Logarithmic area-under-the-curve}

\item{x1}{Numeric indicating the minimal tested logarithmic concentration}

\item{x2}{Numeric indicating the maximal tested logarithmic concentration}

\item{log_extra}{Integer indicating the extra log orders around the measured concentration range beyond which the logEC50 is clipped}
}
\value{
Normalized EC50 for the logarithmic dose response curve
}
\description{
\code{calculate_norm_logEC50()} computes the normalized (feature scaled) logEC50.
The minimal logEC50 is defined as \code{x1 - log_extra} and the maximal logEC50 as \code{x2 + log_extra},
where \code{x1} and \code{x2} are the lowest and highest tested logarithmic concentration.
}
\concept{scoring}
