% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_percentile}
\alias{calculate_percentile}
\title{Calculate percentiles of fit parameters}
\usage{
calculate_percentile(tbl, ..., probs = c(0.1, 0.33, 0.5, 0.66, 0.9))
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} with at least the following columns
\itemize{
\item drug_id
\item and the metrics selected in ...
}}

\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Columns names of fitting parameters to calculate a percentile of}

\item{probs}{Numeric vector with probabilities to calculate percentiles for}
}
\value{
A \link[tibble]{tibble} including percentile and corresponding boolean columns indicating
whether the observation is below the xxth-percentile.
}
\description{
Calculate percentile columns for the fitting parameters (logIC50, logAUC, etc.) as well as boolean columns
indicating whether the observation is below the xxth-percentile. Percentiles are calculated based on the cohort in \code{tbl}
excluding assays that have failed QC.
}
\concept{scoring}
