% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_proliferation}
\alias{calculate_proliferation}
\title{Calculate proliferation index}
\usage{
calculate_proliferation(tbl, con, readout_id = c(1, 5))
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} with at least the following columns:
\itemize{
\item sid
\item assay_id
\item readout_id (optional)
}}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}

\item{readout_id}{Vector of integers specifying the readout IDs of which the proliferation index is calculated}
}
\value{
Numeric specifying the proliferation ratio for all assays/samples
}
\description{
The proliferation capacity is assessed by comparing the number of cancer cells at 96h and 18h (right before drug addition).
A value above 1 means that the cells have proliferated on average while a value below 1 means there is loss of cells viability due to drug-independent cell death.
}
\note{
The response must be of the same \code{readout_id} at 18h and at the endpoint.
}
\concept{scoring}
