% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_relative_logEC}
\alias{calculate_relative_logEC}
\title{Calculate relative logEC}
\usage{
calculate_relative_logEC(effect, b, c, d, logEC50)
}
\arguments{
\item{effect}{Percent relative effect (e.g. 50 or 20)}

\item{b}{Slope parameter}

\item{c}{Lower asymptote}

\item{d}{Upper asymptote}

\item{logEC50}{Logarithmic half-maximum effective concentration (relative EC50)}
}
\value{
Concentration at which the given response is observed
}
\description{
\code{calculate_relative_logEC} computes the relative effect concentration (e.g. relative logEC50 or logEC20) for the LL.2, LL.3 or LL.4 model.
The relative logEC50 is the concentration at which the effect is half-maximal relative to the
maximal effect (Emax). The Emax relates to a residual cell population not affected by the drug.
Note: logEC20 -> 20\% relative effect = 20\% relative inhibition = 80\% relative viability
}
\note{
This is equivalent to \code{drc::ED(mdl, effect/100, type="relative")} where \code{mdl} is a \code{drc::drm()} object
}
\concept{scoring}
