% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_sdss}
\alias{calculate_sdss}
\title{Calculate the differential drug sensitivity score (DSS)}
\usage{
calculate_sdss(dss, mean_dss_ctrl)
}
\arguments{
\item{dss}{Drug sensitivity score}

\item{mean_dss_ctrl}{Average drug sensitivity score of the control cohort}
}
\value{
Differential drug sensitivity score
}
\description{
Reference the DSS to the mean of a (healthy) control cohort
}
\concept{scoring}
