% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_variance_score}
\alias{calculate_variance_score}
\title{Patient fingerprint variance score}
\usage{
calculate_variance_score(tbl, ...)
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} with \code{assay_id}, \code{sid} and \verb{sd_<metrics>} columns}

\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Columns names of fitting parameters to calculate a percentile of}
}
\value{
A \link[tibble]{tibble} including additional columns \verb{var_rank_<metric>} for each metric
}
\description{
Computes a patient sensitivity score by weighting the drug fingerprint with
the cohort variance of the drug.
}
\concept{scoring}
