% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_zscore}
\alias{calculate_zscore}
\title{Calculation z-score of fit parameters}
\usage{
calculate_zscore(tbl, ..., by = drug_id)
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} with at least the following columns:
\itemize{
\item drug_id
\item and the metrics selected in ...
}}

\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Columns names of parameters to calculate a z-score of}

\item{by}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Column to group by (e.g. drug_id or sid)}
}
\value{
A \link[tibble]{tibble} including z-score columns
}
\description{
Calculate z-scores of the fitting parameters (logIC50, logAUC, etc.).
The z-score is calculated based on the cohort in \code{tbl} excluding assays that have failed QC.
}
\concept{scoring}
