% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{calculate_zscore_from_stats}
\alias{calculate_zscore_from_stats}
\title{Calculation z-score with cohort statistics}
\usage{
calculate_zscore_from_stats(tbl, con, cohort_id, ...)
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} with at least the following columns:
\itemize{
\item drug_id
\item and the metrics selected in ...
}}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}

\item{cohort_id}{Integer specifying the cohort_id to use for the z-score calculation}

\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Columns names of parameters to calculate a z-score of}
}
\value{
A \link[tibble]{tibble} including z-score columns
}
\description{
Calculate z-scores of the fitting parameters (logIC50, logAUC, etc.).
The z-score is calculated based on the \code{cohort_stats} table
}
\concept{scoring}
