% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{cli_fit}
\alias{cli_fit}
\title{CLI for curve fitting}
\usage{
cli_fit(
  tbl,
  con,
  fixed = c(NA, NA, 1, NA, 1),
  lowerl = c(0.01, 0, 1e-04),
  upperl = c(2.5, 1, 1e+06),
  start = NULL,
  show_coefficients = FALSE
)
}
\arguments{
\item{tbl}{A tibble with at least the columns:
\itemize{
\item plate_id
\item sid
\item drug_id
\item concentration
\item readout_id
\item norm_response
}}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}

\item{fixed}{A numeric vector specifying the values of the parameters (b, c, d, e, f) that should be fixed in the drc model. Variable parameters should be specified as NA.}

\item{lowerl}{Lower boundaries of the variable parameters}

\item{upperl}{Upper boundaries of the variable parameters}

\item{start}{Starting values of the variable parameters (if NULL the model chooses starting values automatically)}

\item{show_coefficients}{Boolean indicating whether to print out fit coefficients}
}
\description{
A command line interface (CLI) for fitting dose response curves. The CLI internally calls \link{fit_wrapper} to fit, extract parameters
and predict dose response curves using the \link[drc]{drm} package. The parameters and fitted curves are then written to the DRP Postgres database.
}
\examples{
tbl <- tibble::tribble(
  ~sid, ~plate_id, ~drug_id, ~concentration, ~norm_response, ~readout_id,
  1, 1, 1, 0.4, 1, 1,
  1, 1, 1, 1.3, 0.75, 1,
  1, 1, 1, 4.5, 0.3, 1,
  1, 1, 1, 15, 0.1, 1,
  1, 1, 1, 50, 0, 1
)
# cli_fit(tbl, con)
}
\seealso{
\link{fit_wrapper} and \link{fit}
}
\concept{fitting}
