% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Establish a connection to a DBMS}
\usage{
connect(dbname = "drp", driver = RPostgres::Postgres(), pool = TRUE, ...)
}
\arguments{
\item{dbname}{Name of the database}

\item{driver}{Database driver to use, either \code{RSQLite::SQLite()} or \code{RPostgres::Postgres()}}

\item{pool}{Boolean indicating whether the package 'pool' (TRUE) or 'DBI' (FALSE) should be used for database connection}

\item{...}{Additional arguments to pass to \link[pool]{dbPool}, i.e. host, user, password, sslmode}
}
\value{
A \link[DBI]{DBIConnection-class} object to a DRP database
}
\description{
\code{connect} is a wrapper around \link[pool]{dbPool} to connect to a SQLite or Postgres database using the specified driver and
additional authentication parameters (\code{host}, \code{user} and \code{password}).
}
\concept{connecting}
