% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_with_config}
\alias{connect_with_config}
\title{Connect to DB using a configuration file}
\usage{
connect_with_config(
  config_name = "default",
  user_file = system.file("config.user.yml", package = PKG_NAME),
  driver = RPostgres::Postgres(),
  cache = FALSE,
  pool = TRUE
)
}
\arguments{
\item{config_name}{Name of the configuration definition}

\item{user_file}{Name of the user configuration file (default: config.user.yml; base configurations are present in config.yml)}

\item{driver}{Database driver to use}

\item{cache}{Boolean indicating whether to use caching with `R.cache``}

\item{pool}{Boolean indicating whether the package 'pool' (TRUE) or 'DBI' (FALSE) should be used for database connection}
}
\value{
A \link[DBI]{DBIConnection-class} object to a DRP database
}
\description{
\code{connect_with_config} uses connection variables defined in a configuration file to
establish a connection to a DBMS.
}
\concept{connecting}
