% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{cor_pmat}
\alias{cor_pmat}
\title{Correlation p-value}
\usage{
cor_pmat(tbl, ...)
}
\arguments{
\item{tbl}{A numeric matrix or tibble}

\item{...}{Additional arguments to be passed to cor.test.}
}
\description{
Compute the matrix of correlation p-values
}
\concept{scoring}
