% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{custom_palettes}
\alias{custom_palettes}
\title{Custom color palettes}
\usage{
custom_palettes()
}
\value{
list of named HEX color vectors
}
\description{
A list of custom color palettes each consisting of a character vector of HEX colors
}
\concept{coloring}
