% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{dbGetInfo}
\alias{dbGetInfo}
\title{Get info on database connection}
\usage{
dbGetInfo(con)
}
\arguments{
\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}
}
\value{
A named list with \code{dbname}, \code{username}, \code{host}, \code{port} and \code{db.version}
as returned by \code{DBI::dbGetInfo()} on a fetched connection
}
\description{
\code{dbGetInfo} returns the name of the database, host and logged in user.
}
\concept{connecting}
