% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{deactivateCache}
\alias{deactivateCache}
\title{Deactivate R.cache}
\usage{
deactivateCache()
}
\description{
Don't use caching and don't write temporary files to \code{R.cache::getCachePath()}.
}
\concept{helper}
