% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{decode}
\alias{decode}
\title{Decode feature columns}
\usage{
decode(tbl, con)
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} with integer feature columns defined in the DRP database.}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}
}
\value{
A \link[tibble]{tibble} with decoded features (character feature columns)
}
\description{
\code{encode} converts integer columns to character according to the definitions in the specified DRP database.
}
\seealso{
\code{\link[=encode]{encode()}}
}
\concept{transforming}
