% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{derive_fit_features}
\alias{derive_fit_features}
\title{Calculate cohort scores}
\usage{
derive_fit_features(tbl, con)
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} with at least the following columns:
\itemize{
\item pid
\item sid
\item drug_id
\item fit parameter columns b,c,d,e
}}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}
}
\value{
A \link[tibble]{tibble} including derivate fit parameters
}
\description{
Calculate z-scores, (percent)-ranks.
}
\concept{scoring}
