% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{derive_parameters}
\alias{derive_parameters}
\title{Calculate derivate fit parameters}
\usage{
derive_parameters(tbl, con)
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} with at least the following columns:
\itemize{
\item pid
\item sid
\item drug_id
\item fit parameter columns b,c,d,e
}}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}
}
\value{
A \link[tibble]{tibble} including derivate fit parameters
}
\description{
Calculate secondary, derivate fit parameters such as logAUC or abs_logEC50
}
\concept{scoring}
