% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{desc_to_id}
\alias{desc_to_id}
\title{Convert description to ID}
\usage{
desc_to_id(
  description,
  ref_tbl,
  con,
  id_col = id,
  description_col = description,
  ignore_error = TRUE
)
}
\arguments{
\item{description}{Character string to convert to ID}

\item{ref_tbl}{Character string of the table name in the DRP database or a reference \link[tibble]{tibble} with the id-description mapping}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}

\item{id_col}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> name of the ID column in the table}

\item{description_col}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> name of the description column in the table}

\item{ignore_error}{Boolean indicating whether to ignore missing descriptions that lead to NAs}
}
\value{
An integer
}
\description{
Get the ID of a reference data record from the description column in the DRP database.
}
\concept{transforming}
