% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{dose_range}
\alias{dose_range}
\title{Get preferred drug dose range}
\usage{
dose_range(drug, con)
}
\arguments{
\item{drug}{Drug name or drug ID}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}
}
\value{
A named list of the minimum and maximum doses
}
\description{
Get preferred minimum and maximum dose of a specific drug
}
\concept{fitting}
