% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{duplicated_rows}
\alias{duplicated_rows}
\title{Identify duplicated rows}
\usage{
duplicated_rows(tbl, id_cols)
}
\arguments{
\item{tbl}{A \link[tibble]{tibble}}

\item{id_cols}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of identifying column(s)}
}
\value{
A \link[tibble]{tibble} of duplicated rows
}
\description{
Return duplicated rows of a tibble based on a set of identifying columns
}
\concept{helper}
