% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{encode}
\alias{encode}
\title{Encode feature columns}
\usage{
encode(tbl, con)
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} with character feature columns defined in the DRP database.}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}
}
\value{
A \link[tibble]{tibble} with encoded features (integer feature columns)
}
\description{
\code{encode} converts character columns to integers according to the definitions in the specified DRP database.
}
\seealso{
\code{\link[=decode]{decode()}}
}
\concept{transforming}
