% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{encode_csv}
\alias{encode_csv}
\title{Encode comma-separated values}
\usage{
encode_csv(tbl, ref_tbl, con)
}
\arguments{
\item{tbl}{A reference \link[tibble]{tibble} with the id-description mapping.}

\item{ref_tbl}{Character string of the table name in the DRP database or a reference \link[tibble]{tibble} with the id-description mapping}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}
}
\value{
A \link[tibble]{tibble} with encoded features (comma-delimited integer vector)
}
\description{
\code{encode_csv} converts a comma-separated string of characters values to a comma-separated string of integers
according to the definitions in the specified DRP database.
}
\seealso{
\code{\link[=encode]{encode()}}, \code{\link[=decode]{decode()}}
}
\concept{transforming}
