% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit a drug response model}
\usage{
fit(
  tbl,
  fixed = c(NA, NA, 1, NA, 1),
  lowerl = c(0.01, 0, 1e-04),
  upperl = c(2.5, 1, 1e+06),
  start = c(1, 0, 100)
)
}
\arguments{
\item{tbl}{A tibble with the columns:
\itemize{
\item concentration
\item norm_response
}}

\item{fixed}{A numeric vector specifying the values of the parameters (b, c, d, e, f) that should be fixed in the drc model. Variable parameters should be specified as NA.}

\item{lowerl}{Lower boundaries of the variable parameters}

\item{upperl}{Upper boundaries of the variable parameters}

\item{start}{Starting values of the variable parameters (if NULL the model chooses starting values automatically)}
}
\value{
A \link[drc]{drm} model
}
\description{
Fit a dose-response model using a \link[drc]{drm} package.
}
\concept{fitting}
