% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_wrapper}
\alias{fit_wrapper}
\title{Curve fitting wrapper}
\usage{
fit_wrapper(
  tbl,
  con,
  cur,
  drug_id,
  plate_id,
  mdl_name,
  fixed,
  lowerl,
  upperl,
  start,
  readout_id,
  sid
)
}
\arguments{
\item{tbl}{A tibble with at least the columns:
\itemize{
\item sid
\item drug_id
\item plate_id
}}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}

\item{cur}{A named vector with user_id and tenant_id}

\item{drug_id}{Integer identifying the drug}

\item{plate_id}{Integer identifying the plate}

\item{mdl_name}{Name of the fitting model}

\item{fixed}{A numeric vector specifying the values of the parameters (b, c, d, e, f; in this order) that should be fixed in the drc model. Variable parameters should be specified as NA.}

\item{lowerl}{Lower boundaries of the variable parameters}

\item{upperl}{Upper boundaries of the variable parameters}

\item{start}{Starting values of the variable parameters (if NULL the model chooses starting values automatically)}

\item{readout_id}{Integer specifying the readout (e.g. 1 = Viable leukemia cells)}

\item{sid}{Integer specifying the sample ID}
}
\description{
Wrapper around the \link{fit} function for carrying out logistic curve fits, extract parameters and predicted curves.
The results are then written to the DRP Postgres database.
}
\concept{fitting}
