% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_configuration}
\alias{get_configuration}
\title{Get configuration settings}
\usage{
get_configuration(
  config_name = "default",
  user_file = system.file("config.user.yml", package = PKG_NAME)
)
}
\arguments{
\item{config_name}{String specifying the configuration name}

\item{user_file}{Path to a user defined configuration file (default: config.user.yml located under /inst)}
}
\value{
A list with the configuration settings
}
\description{
Load configuration settings from \verb{inst/config.yml"} file and
optionally from a user-defined configuration file (by default \code{inst/config.user.yml}).
}
\concept{connecting}
