% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_file_type}
\alias{get_file_type}
\title{Get file type}
\usage{
get_file_type(filename)
}
\arguments{
\item{filename}{Character string specifying the filename}
}
\value{
Character string specifying the file type (\code{img}, \code{doc}, \code{NA})
}
\description{
Get file type based on file extension.
}
\concept{helper}
