% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{get_hex_colors}
\alias{get_hex_colors}
\title{HEX color from a colormap}
\usage{
get_hex_colors(name, n = 3, n_max = NULL)
}
\arguments{
\item{name}{Character string specifying the colormap. Can be registered colormap from \link[RColorBrewer]{RColorBrewer}
or a palette fom the \code{palettes} list defined in the drpr package. Reverse the colormap by appending the suffix \verb{_r} to the name.}

\item{n}{Number of colors to extract.}

\item{n_max}{Maximum number of colors to use from the colormap. If \code{NULL} defaults to all colors}
}
\value{
Character vector with HEX color codes
}
\description{
Extract HEX color codes from a named colormap.
}
\examples{
get_hex_colors("Set2")
get_hex_colors("orange_violet")
}
\concept{coloring}
