% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{id_to_desc}
\alias{id_to_desc}
\title{Convert ID to description}
\usage{
id_to_desc(
  id,
  ref_tbl,
  con,
  id_col = id,
  description_col = description,
  ignore_error = TRUE
)
}
\arguments{
\item{id}{integer to convert to description}

\item{ref_tbl}{Character string of the table name in the DRP database or a reference \link[tibble]{tibble} with the id-description mapping}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}

\item{id_col}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> name of the ID column in the table}

\item{description_col}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> name of the description column in the table}

\item{ignore_error}{Boolean indicating whether to ignore missing IDs that lead to NAs}
}
\value{
A character string
}
\description{
Get the description of a reference data record from the ID column in the DRP database.
}
\concept{transforming}
