% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{initialize_parameters}
\alias{initialize_parameters}
\title{Initialize fit parameters}
\usage{
initialize_parameters()
}
\value{
List of fit parameters for the different models
}
\description{
Provides default start parameters and boundaries for the LL.2, LL.3, LL.4 and LL.5 models.
}
\concept{fitting}
