% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{inverse_ll4}
\alias{inverse_ll4}
\title{Invert the log-logistic curve}
\usage{
inverse_ll4(y, b, c, d, logEC50)
}
\arguments{
\item{y}{Drug response}

\item{b}{Slope parameter}

\item{c}{Lower asymptote}

\item{d}{Upper asymptote}

\item{logEC50}{Logarithmic half-maximum effective concentration (relative EC50)}
}
\value{
Concentration at which the given response is observed
}
\description{
\code{inverse_ll4} calculates the concentration (x) for a given (absolute) response (y)
by solving the four-parameter log-logistic function for x.
}
\seealso{
\code{\link[=calculate_absolute_logEC]{calculate_absolute_logEC()}}
}
\concept{fitting}
