% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{lighten_color}
\alias{lighten_color}
\title{Make color lighter}
\usage{
lighten_color(colors, factor = 0.75)
}
\arguments{
\item{colors}{Character vector of colors either as HEX codes or built-in color names}

\item{factor}{Numeric factor by which to lighten the colors.}
}
\value{
Character vector of lightened colors
}
\description{
Make colors lighter by a given factor.
}
\concept{coloring}
