% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{linspace}
\alias{linspace}
\title{Equally linearly-spaced numbers}
\usage{
linspace(x1, x2, n = 100)
}
\arguments{
\item{x1}{Minimum number}

\item{x2}{Maximum number}

\item{n}{Length of the resulting vector}
}
\value{
Numeric vector of length n
}
\description{
Generate a numeric vector of equally spaced numbers
between a minimum and maximum on a linear scale
}
\concept{helper}
