% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{ll4}
\alias{ll4}
\title{Four-parameter log-Logistic model}
\usage{
ll4(x, b, c, d, logEC50)
}
\arguments{
\item{x}{Logarithmic concentration}

\item{b}{Inverse slope parameter}

\item{c}{Lower asymptote}

\item{d}{Upper asymptote}

\item{logEC50}{Logarithmic half-maximum effective concentration (relative EC50)}
}
\value{
Predicted effect
}
\description{
The log-logistic four-parameter model (ll.4) as implemented by the \link[drc]{drm} package
and decribed from Ritz et al. PLOS ONE 2015.
}
\concept{fitting}
