% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{negative_control_stat}
\alias{negative_control_stat}
\title{Calculate negative control statistics}
\usage{
negative_control_stat(tbl, negative_control = "DMSO", FUN = mean)
}
\arguments{
\item{tbl}{A tibble with at least the following columns present:
\itemize{
\item drug
\item response
}}

\item{negative_control}{A string defining the substance to use as negative control}

\item{FUN}{An averaging function, i.e. mean or median}
}
\value{
A dbl of the average negative control response
}
\description{
Calculate the mean or median of the DRP negative control.
}
\concept{transforming}
