% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{next_id}
\alias{next_id}
\title{Get next ID}
\usage{
next_id(con, tbl, pk = NULL, exclude = NULL, named = FALSE)
}
\arguments{
\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}

\item{tbl}{A character string specifying the database table name}

\item{pk}{A character string specifying the primary key column name (if \code{NULL} look up the pk in the database)}

\item{exclude}{A numeric vector of IDs in the primary key to exclude from the search}

\item{named}{A logical indicating whether to return a named vector with the primary key name}
}
\value{
An integer or named vector specifying the next free ID
}
\description{
Get the next ID from the primary key column of a database table
}
\concept{helper}
