% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{normalize_measurements}
\alias{normalize_measurements}
\title{Normalize measurements}
\usage{
normalize_measurements(con, plate_id)
}
\arguments{
\item{con}{A \link[DBI]{DBIConnection-class} object, as returned by \link[DBI]{dbConnect}}

\item{plate_id}{Numeric indicating which plate should be normalized.}
}
\description{
Normalize measurements to the DMSO control.
}
\concept{transforming}
