% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{normalize_response}
\alias{normalize_response}
\title{Normalize drug responses}
\usage{
normalize_response(
  plate,
  negative_control = "DMSO",
  negative_control_mean = NULL,
  remove_control = FALSE
)
}
\arguments{
\item{plate}{A tibble with at least the following columns:
\itemize{
\item drug
\item response
}}

\item{negative_control}{A character vector describing the treatment (drug) used as a negative control (e.g. DMSO)}

\item{negative_control_mean}{A manually pre-calculated, mean DMSO response (e.g. if no DMSO was measured on the current plate).
Overrules any DMSO observations present on the plate.}

\item{remove_control}{Boolean indicating whether to remove DMSO observations from the resulting tibble}
}
\value{
A \link[tibble]{tibble} with normalized responses added
}
\description{
\code{normalize_response} divides each readout observation by the mean DMSO response and adds a column \code{norm_response} to the tibble
}
\concept{transforming}
