% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{oauth}
\alias{oauth}
\title{Generate an OAuth client}
\usage{
oauth(config_name = "api")
}
\arguments{
\item{config_name}{Character string referring to the name of the configuration definition}
}
\value{
A list with a \link[httr2]{oauth_client}, the auth_url, the redirect_uri and the scope
}
\description{
Read OAuth client ID, secret, and token/authentication URL from a (user) config file
and return a \link[httr2]{oauth_client}.
}
\concept{connecting}
