% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot drug distribution}
\usage{
plot_distribution(
  con,
  .jitter_mask = TRUE,
  .point_mask = FALSE,
  .drug_mask = TRUE,
  x = drug_name,
  metric = norm_logAUC,
  color_fill = abs_logEC50,
  readout_id = c(1, 5),
  bg_readout_id = c(1, 5),
  sort_by = NULL,
  label = lknumber,
  selected_labels = NULL,
  palette = NULL,
  color_fill_name = NULL,
  add_boxplot = FALSE,
  add_line = FALSE,
  quantiles = FALSE,
  lower_quantile = p10,
  upper_quantile = p90,
  shape = 16,
  size = 2,
  alpha = 0.5,
  limits = NULL,
  midpoint = NULL,
  ylimits = NULL,
  lump_n = NULL,
  drop_l = NULL,
  na_as_other = FALSE,
  n_levels = NULL,
  x_var = 0,
  bg_x_var = 0.1,
  bg_color = NULL,
  seed = 123,
  comparisons = NULL,
  test = "wilcox.test",
  paired_test = FALSE,
  stat_compare_label = "p.signif",
  stat_compare_label.y = NULL,
  progressbar = TRUE,
  save = FALSE,
  filename = NULL,
  width = 8,
  height = 3.5
)
}
\arguments{
\item{con}{\link[DBI]{DBIConnection-class} object to a DRP database}

\item{.jitter_mask}{<\code{\link[rlang:args_data_masking]{data-masking}}> expression to select the samples/assays for the \code{geom_jitter()}}

\item{.point_mask}{<\code{\link[rlang:args_data_masking]{data-masking}}> expression to select the samples/assays for the \code{geom_point()}}

\item{.drug_mask}{<\code{\link[rlang:args_data_masking]{data-masking}}> expression to select the drugs}

\item{x}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of the factor to use as the independent variable (x-axis)}

\item{metric}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of the DRP metric (continuous) use as the dependent variable (y-axis)}

\item{color_fill}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of the column to be used for color and fill}

\item{readout_id}{Integer or vector of integers specifying the id(s) of the readout in the point plot}

\item{bg_readout_id}{Integer or vector of integers specifying the id(s) of the readout in the jitter plot}

\item{sort_by}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of the column to sort the x-axis by (e.g. \code{norm_logAUC}). Invert the order with \verb{dplyr::desc(<sort_by>)}.}

\item{label}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of a column to be used for labeling each drug curve}

\item{selected_labels}{<\code{\link[rlang:args_data_masking]{data-masking}}> expression returning a logical vector and selecting the labels to be displayed.
Example: for \code{label = sid} set for instance \code{selected_labels = sid \%in\% c(1, 2)} to exclusively label these two SIDs}

\item{palette}{Character vector specifying the color palette for color and fill (discrete or continuous)}

\item{color_fill_name}{Character string specifying an alternative name for the color/fill legend}

\item{add_boxplot}{Boolean specifying whether to add boxplots}

\item{add_line}{Boolean specifying whether to add lines connecting the \code{geom_point}}

\item{quantiles}{Boolean indicating whether to draw boxplot from quantiles of \code{cohort_stats}}

\item{lower_quantile}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of the lower percentile column in the \code{cohort_stats} table specifying the min boxplot whisker. To hide whiskers set \code{lower_quantile = p25}}

\item{upper_quantile}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of the upper percentile column in the \code{cohort_stats} table specifying the maximum boxplot whisker. To hide whiskers set \code{upper_quantile = p75}}

\item{shape}{Integer specifying the \code{geom_point} type}

\item{size}{Integer specifying the size of the jitter dots}

\item{alpha}{Numeric specifying the transparency of the jitter dots}

\item{limits}{Numeric vector of lower and upper limits of the color/fill scale}

\item{midpoint}{Numeric specifying the midpoint of colorscale}

\item{ylimits}{Numeric vector of the lower and upper limits on the y-axis}

\item{lump_n}{Named vector of integers specifying the maximum number of levels for each feature (including the lumped "Other" category),
e.g. c(diagnosis = 3)}

\item{drop_l}{Named list of character vectors specifying the levels to drop for each feature}

\item{na_as_other}{Boolean indicating whether to lump NA values into the "Other" category}

\item{n_levels}{Integer indicating the maximum number of levels to display on the x-axis}

\item{x_var}{Numeric indicating the amount of random variation to add to the points on the x-axis}

\item{bg_x_var}{Numeric indicating the amount of random variation to add to the background jitter points on the x-axis}

\item{bg_color}{Character specifying a uniform color for the background jitter points}

\item{seed}{Integer specifying the seed for the random number generator}

\item{comparisons}{A list of length-2 vectors. The entries in the vector are either the names of two values on the x-axis or the two integers that correspond to the index of the groups of interest, to be compared.}

\item{test}{Character string describing the statistical test to use for the comparisons. Allowed values include "wilcox.test" (Wilcoxon rank sum test) and "t.test" (Student's t-test)}

\item{paired_test}{Boolean indicating whether to use a paired test}

\item{stat_compare_label}{Character string specifying label type. Allowed values include "p.signif" (shows the significance levels) and "p.format" (shows the formatted p-value)}

\item{stat_compare_label.y}{Numeric indicating the y-axis position of the p-value label}

\item{progressbar}{Boolean indicating whether to show a progressbar}

\item{save}{Boolean indicating whether to write the plot to disk}

\item{filename}{Character string specifying the filename}

\item{width}{Numeric specifying the width of the plot}

\item{height}{Numeric specifying the height of the plot}
}
\value{
A \link[ggplot2]{ggplot} object
}
\description{
Distribution dot plot of the best drugs as a function of an ex vivo scoring metric (logEC50, logAUC, DSS).
The dots can be colored by an second independent metric (e.g. top 10\% outliers, diagnosis, disease stage etc.).
The distribution plot can be combined with other geoms, see examples.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
plot_distribution(con,
  .drug_mask = functional_class == "MEK inhibitor",
  color_fill = TP53,
  add_boxplot = TRUE
)
\dontshow{\}) # examplesIf}
}
\concept{plotting}
