% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_plate}
\alias{plot_plate}
\title{Draw a multi-well plate}
\usage{
plot_plate(
  plate,
  color,
  alpha,
  n_wells = NULL,
  row = row,
  col = col,
  label = NULL,
  label_size = 2,
  text = NULL,
  palette = "hue12",
  limits = NULL,
  midpoint = NULL,
  transform = "identity",
  show_legend = TRUE,
  save = FALSE,
  filename = NULL,
  width = 6,
  height = 4
)
}
\arguments{
\item{plate}{A \link[tibble]{tibble} with at least the following columns:
\itemize{
\item row
\item col
and usually either
\item drug
\item concentration
or
\item response/norm_response
}}

\item{color}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of the column to map the color}

\item{alpha}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of the column to map the opacity}

\item{n_wells}{Integer specifying the number of wells of the plate, if \code{NULL} derive the well number from \code{plate}}

\item{row}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of the column to identify the plate rows}

\item{col}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of the column to identify the plate columns}

\item{label}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of a column to be used for labeling the well}

\item{label_size}{Integer specifying the size of the label}

\item{text}{Name of the column to be used for labeling the tooltips in the interactive plot}

\item{palette}{A character vector specifying the color palette (discrete or continuous)}

\item{limits}{A numeric vector of lower and upper limits of the color/fill scale}

\item{midpoint}{Numeric specifying the midpoint of colorscale}

\item{transform}{For continuous scales the name of the transformation object (e.g. "log10") or the object itself ("identity")}

\item{show_legend}{Boolean indicating whether to show the legend}

\item{save}{Boolean indicating whether to write the plot to disk}

\item{filename}{Character string specifying the filename}

\item{width}{Numeric specifying the width of the plot}

\item{height}{Numeric specifying the height of the plot}
}
\value{
A \link[ggplot2]{ggplot} object
}
\description{
Draw a multi-well plate
}
\concept{plotting}
