% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_profile}
\alias{plot_profile}
\title{Plot drug response profiles}
\usage{
plot_profile(
  con,
  .drug_mask,
  .fg_sample_mask = FALSE,
  .bg_cohort_mask = TRUE,
  color_fill = abs_logEC50,
  label = lknumber,
  size = NULL,
  readout_id = c(1, 5),
  bg_readout_id = c(1, 5),
  transform = "identity",
  palette = NULL,
  limits = NULL,
  midpoint = NULL,
  quantiles = FALSE,
  lower_quantile = p10,
  upper_quantile = p90,
  sort_by = NULL,
  n_drugs = NULL,
  drop_missing_drugs = FALSE,
  lump_n = NULL,
  drop_l = NULL,
  na_as_other = FALSE,
  ylimits = c(0, 1.05),
  xlimits = NULL,
  save = FALSE,
  filename = NULL,
  color_fill_name = NULL,
  bg_color = NULL,
  show_na = TRUE,
  progressbar = TRUE,
  facet_samples = TRUE,
  nrow = NULL,
  ncol = NULL,
  width = 3.6,
  height = 2.5
)
}
\arguments{
\item{con}{\link[DBI]{DBIConnection-class} object to a DRP database}

\item{.drug_mask}{<\code{\link[rlang:args_data_masking]{data-masking}}> expression returning a logical vector and selecting the drugs to be displayed}

\item{.fg_sample_mask}{<\code{\link[rlang:args_data_masking]{data-masking}}> expression returning a logical vector and
selecting the samples displayed in the foreground based on the columns in plate_content_view}

\item{.bg_cohort_mask}{<\code{\link[rlang:args_data_masking]{data-masking}}> expression to select the samples displayed in the background}

\item{color_fill}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of the column to be used for color and fill}

\item{label}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of a column to be used for labeling each drug curve}

\item{size}{Numeric specifying the line size}

\item{readout_id}{Integer or vector of integers specifying the id(s) of the readout in the foreground}

\item{bg_readout_id}{Integer or vector of integers specifying the id(s) of the readout in the background}

\item{transform}{For continuous scales the name of the transformation object (e.g. "log10") or the object itself ("identity")}

\item{palette}{A character vector specifying the color palette (discrete or continuous)}

\item{limits}{A numeric vector of lower and upper limits of the color/fill scale}

\item{midpoint}{Numeric specifying the midpoint of colorscale}

\item{quantiles}{Boolean indicating whether to draw a percentile interval instead of individual background curves}

\item{lower_quantile}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of the lower percentile column in the \code{cohort_stats} table}

\item{upper_quantile}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of the upper percentile column in the \code{cohort_stats} table}

\item{sort_by}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Name of the column to be used for sorting facets}

\item{n_drugs}{Integer specifying the maximal number of drugs}

\item{drop_missing_drugs}{Boolean indicating whether to drop drugs that are missing in the \code{.fg_sample_mask}}

\item{lump_n}{Named vector of integers specifying the maximum number of levels for each feature (including the lumped "Other" category),
e.g. c(diagnosis = 3)}

\item{drop_l}{Named list of character vectors specifying the levels to drop for each feature}

\item{na_as_other}{Boolean indicating whether to lump NA values into the "Other" category}

\item{ylimits}{Numeric vector of specifying the y-axis limits}

\item{xlimits}{Numeric vector specifying the x-axis limits}

\item{save}{Boolean indicating whether to write the plot to disk}

\item{filename}{Character string specifying the filename}

\item{color_fill_name}{Character string specifying an alternative name for the color/fill legend}

\item{bg_color}{Character specifying a uniform color for the background fit lines}

\item{show_na}{Boolean indicating whether to translate NA values}

\item{progressbar}{Boolean indicating whether to show a progressbar}

\item{facet_samples}{Boolean indicating whether to separate samples into individual facets}

\item{nrow}{Integer indicating the number of rows to use for facetting}

\item{ncol}{Integer indicating the number of columns to use for facetting}

\item{width}{Numeric specifying the width of the plot}

\item{height}{Numeric specifying the height of the plot}
}
\value{
\link[ggplot2]{ggplot} object
}
\description{
Plotting drug responses as a function of concentration.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
plot_profile(con,
  .drug_mask = drug_name \%in\% c("Venetoclax", "Navitoclax"),
  .fg_sample_mask = lknumber == "LK2000.001",
  color_fill = norm_logAUC,
  palette = "orange_blue",
  color_fill_name = "logAUC"
)
\dontshow{\}) # examplesIf}
}
\concept{plotting}
