% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{predict_drc}
\alias{predict_drc}
\title{Predict drug responses}
\usage{
predict_drc(mdl, min, max, n = 25, interval = "confidence")
}
\arguments{
\item{mdl}{A model of the \link[drc]{drm} class}

\item{min}{The minimum drug concentration}

\item{max}{The maximum drug concentration}

\item{n}{Integer specifying the number of data points to predict}

\item{interval}{The 95\% confidence intervals around the mean (interval = "confidence") or the 95\% prediction interval (interval = "prediction").}
}
\value{
A \link[tibble]{tibble} with predictions and confidence intervals in the following columns:
\itemize{
\item Concentration
\item Prediction
\item Lower
\item Upper
}
}
\description{
Predict drug responses for a n-series of evenly spaced concentrations on a logarithmic scale given a minimum and maximum dose.
}
\concept{fitting}
