% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{reference}
\alias{reference}
\title{Get reference data}
\usage{
reference(con, with_synonyms = TRUE)
}
\arguments{
\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}

\item{with_synonyms}{Boolean indicating whether to include drug synonyms}
}
\value{
A list with reference \link[tibble]{tibble}
}
\description{
Return the list of all reference tables in the DRP database.
}
\concept{transforming}
