% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{relabel_fit_parameters}
\alias{relabel_fit_parameters}
\title{Relabel fit parameters columns}
\usage{
relabel_fit_parameters(tbl, data_type = "viability")
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} with columns b,c,d,e}

\item{data_type}{A character string describing if the \link[drc]{drm} parameters are determined by fitting a "inhibition" or "viability" curve}
}
\value{
A \link[tibble]{tibble} with EC50, inverse_slope (N_inv), min_inhibition (I_min), min_viability (I_max), min_viability (V_min) and max_viability (V_max) columns added
}
\description{
Create a copy of the default fit parameter columns (b,c,d,e) and make their names interpretable as pertaining to an inhibition and viability dose response model
\itemize{
\item \code{b} is the inverse slope
\item \code{c} is the lower asymptote corresponding to the minimum inhibition in an inhibition model or the minimum viability in an viability model
\item \code{d} is the upper asymptote corresponding to the maximum inhibition in an inhibition model or the maximum viability in a viability model
\item \code{e} is the EC50
}
}
\seealso{
\code{\link[=extract_fit_parameters]{extract_fit_parameters()}}
}
\concept{fitting}
