% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{request}
\alias{request}
\title{Httr2 request wrapper}
\usage{
request(uri, method = "get", body = NULL, oa = NULL)
}
\arguments{
\item{uri}{A string specifying the URI endpoint. Supports \link[glue]{glue} syntax.}

\item{method}{A string specifying the HTTP method ("post" or "put")}

\item{body}{A list or \link[tibble]{tibble} to send as the body}

\item{oa}{An OAuth list return by \code{\link[=oauth]{oauth()}} with a \link[httr2]{oauth_client}, the auth_url, the redirect_uri and the scope}
}
\value{
An atomic vector, list or \code{NULL}
}
\description{
A helper function to send a \link[httr2]{httr2} request to the DRP API.
}
\concept{connecting}
