% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{runAPI}
\alias{runAPI}
\title{Run Plumber API}
\usage{
runAPI(
  config_name = "default",
  port = 3840,
  background = FALSE,
  pkg_path = "."
)
}
\arguments{
\item{config_name}{Character string referring to the name of the configuration definition}

\item{port}{Integer indicating the port of the API}

\item{background}{Boolean indicating whether to run the Shiny app in the background and free the R console.}

\item{pkg_path}{A character string specifying the (relative) path to the package}
}
\description{
Run a local Plumber API server with a predefined configuration to connect to a local or remote Postgres database.
}
\concept{connecting}
