% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scale_color_fill_custom}
\alias{scale_color_fill_custom}
\title{Add custom color and fill scale}
\usage{
scale_color_fill_custom(
  data,
  color_fill,
  aesthetics = c("color", "fill"),
  palette = NULL,
  limits = NULL,
  midpoint = NULL,
  transform = "identity",
  color_fill_name = NULL,
  show_na = TRUE,
  na.value = "gray"
)
}
\arguments{
\item{data}{\link[tibble]{tibble} of the dataset passed to \link[ggplot2]{ggplot}}

\item{color_fill}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> color and/or fill aesthetics}

\item{aesthetics}{Character string or vector specifying the names of the aesthetics to use for the scale.
This is useful to a set the same color settings to multiple aesthetics simultaneously.}

\item{palette}{Color palette specified by either:
\itemize{
\item a character string specifying a colormap from colorbrewer
\item a character string from the global variable "palettes"
\item a character string specifying a colormap from paletteer::paletteer_d()
\item a character vector of HEX colors starting with "#"
\item a character vector of built-in colors specifying
}}

\item{limits}{Numeric vector specifying the limits of the color scale}

\item{midpoint}{Numeric specifying the midpoint of colorscale}

\item{transform}{Character string specifying the transformation to be applied to the color scale, e.g. "log10"}

\item{color_fill_name}{Character string specifying the name of the color scale. If NULL use the name of the color_fill tidy-selection}

\item{show_na}{Boolean indicating whether to show NA in the legend}

\item{na.value}{Character indicating the color to use for NA values}
}
\value{
List of ggplot2 color and/or fill scale(s)
}
\description{
Add a custom color and/or fill scale depending on the type of input data
}
\concept{coloring}
