% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{select_from_pg}
\alias{select_from_pg}
\title{Select tibble columns from Postgres}
\usage{
select_from_pg(tbl, name, con)
}
\arguments{
\item{tbl}{A \link[tibble]{tibble} to select columns from}

\item{name}{Character string specifying the name of the Postgres table}

\item{con}{A \link[DBI]{DBIConnection-class} object to a DRP database}
}
\value{
A \link[tibble]{tibble} with the selected columns
}
\description{
Selects columns of a tibble based on the variables present in a given Postgres table.
}
\concept{transforming}
