% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{statusCache}
\alias{statusCache}
\title{Status of R.cache}
\usage{
statusCache()
}
\description{
Enquire the status of caching and return the number of key files from \code{R.cache::getCachePath()}.
}
\concept{helper}
