% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_drpr}
\alias{theme_drpr}
\title{Add drpr plot theme to ggplot}
\usage{
theme_drpr()
}
\description{
Theme based on \link[ggpubr]{theme_pubr} with rectangular borders and legend positioned on the right.
}
\concept{plotting}
