% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{validate_layout}
\alias{validate_layout}
\title{Validate layout}
\usage{
validate_layout(layout)
}
\arguments{
\item{layout}{A \link[tibble]{tibble} to be checked for compliance with the layout header requirements}
}
\value{
Boolean
}
\description{
Validate headers of a layout tibble for completeness
}
\concept{io}
