% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{validate_measurement}
\alias{validate_measurement}
\title{Validate measurement}
\usage{
validate_measurement(measurement)
}
\arguments{
\item{measurement}{A \link[tibble]{tibble} to be checked for compliance with the measurement header requirements}
}
\value{
Boolean
}
\description{
Validate headers of a measurement tibble for completeness
}
\concept{io}
